MLP/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
astroid=2.4.2=py37_0
blas=1.0=mkl
ca-certificates=2020.6.24=0
certifi=2020.6.20=py37_0
colorama=0.4.3=py_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py37_0
freetype=2.10.2=hd328e21_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
intel-openmp=2020.1=216
isort=4.3.21=py37_0
joblib=0.16.0=py_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.2.0=py37h74a9793_0
lazy-object-proxy=1.4.3=py37he774522_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_0
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py37h64f37c6_0
mccabe=0.6.1=py37_1
mkl=2020.1=216
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.1.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
ninja=1.9.0=py37h74a9793_0
numpy=1.18.5=py37h6530119_0
numpy-base=1.18.5=py37hc3f5095_0
olefile=0.46=py37_0
openssl=1.1.1g=he774522_0
pandas=1.0.5=py37h47e9c7a_0
pillow=7.2.0=py37hcc1f983_0
pip=20.1.1=py37_1
pylint=2.5.3=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
python=3.7.7=h81c818b_4
python-dateutil=2.8.1=py_0
pytorch=1.5.1=py3.7_cuda102_cudnn7_0
pytz=2020.1=py_0
qt=5.9.7=vc14h73c81de_0
scikit-learn=0.23.1=py37h25d0782_0
scipy=1.5.0=py37h9439919_0
setuptools=49.1.0=py37_0
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
sqlite=3.32.3=h2a8f88b_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
toml=0.10.1=py_0
torchvision=0.6.1=py37_cu102
tornado=6.0.4=py37he774522_1
typed-ast=1.4.1=py37he774522_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.34.2=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=py37he774522_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.4=ha9fde0e_3
